requirements.txt
asgiref==3.7.2
astroid==2.15.6
bcrypt==4.0.1
dill==0.3.6
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==20.1.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
passlib==1.7.4
platformdirs==3.9.1
psycopg2-binary==2.9.7
PyJWT==1.7.1
pylint==2.17.5
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.7.1
uWSGI==2.0.21
wrapt==1.15.0
